+++ frontend/src/metabase/entities/schemas.js
@@ -2,14 +2,14 @@ import { updateIn } from "icepick";
 import { createEntity } from "metabase/lib/entities";
 
 import { GET } from "metabase/lib/api";
-import {
-  getCollectionVirtualSchemaId,
-  getQuestionVirtualTableId,
-} from "metabase/lib/saved-questions";
 import { generateSchemaId, parseSchemaId } from "metabase/lib/schema";
 
 import { SchemaSchema } from "metabase/schema";
 import Questions from "metabase/entities/questions";
+import {
+  getCollectionVirtualSchemaId,
+  getQuestionVirtualTableId,
+} from "metabase-lib/lib/metadata/utils/saved-questions";
 
 // This is a weird entity because we don't have actual schema objects
