--- frontend/src/metabase/lib/saved-questions/index.js
@@ -1,1 +1,0 @@
-export * from "./saved-questions";
