+++ frontend/src/metabase/query_builder/components/DatasetEditor/DatasetFieldMetadataSidebar/MappedFieldPicker/MappedFieldPicker.tsx
@@ -1,10 +1,10 @@
 import React, { useCallback, useRef } from "react";
 import { t } from "ttag";
 
-import { isVirtualCardId } from "metabase/lib/saved-questions";
 import { SchemaTableAndFieldDataSelector } from "metabase/query_builder/components/DataSelector";
 
 import Fields from "metabase/entities/fields";
+import { isVirtualCardId } from "metabase-lib/lib/metadata/utils/saved-questions";
 import Field from "metabase-lib/lib/metadata/Field";
 
 import { StyledSelectButton } from "./MappedFieldPicker.styled";
